pyproject.toml
[project]
name = "datasetconverter"
version = "0.1.0"
description = "This project consist of a library and a CLI for converting datasets between annotation formats."
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Guillermo Cousido Martínez", email = "guillermo.cousido@gmail.com" }
]
license = { text = "MIT" }
keywords = ["dataset", "converter", "computer vision", "cli", "yolo", "coco", "pascal voc"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
dependencies = [
    "click>=8.1.8",
    "pillow>=11.2.1",
]

[project.urls]
Homepage = "https://github.com/GCousido/TFG-DatasetConverter"
Documentation = "https://github.com/GCousido/TFG-DatasetConverter#readme"
Repository = "https://github.com/GCousido/TFG-DatasetConverter"
Issues = "https://github.com/GCousido/TFG-DatasetConverter/issues"
Changelog = "https://github.com/GCousido/TFG-DatasetConverter/releases"

[project.scripts]
dconverter = "datasetconverter.cli.main:dconverter"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["datasetconverter/tests"]